refactor(coin-bitcoin): movez cash-specific logic into chain adapters
