+++ libs/coin-modules/coin-bitcoin/src/chain-adapters/default.ts
@@ -1,0 +1,9 @@
+import type { ChainAdapter } from "./types";
+
+/**
+ * No-op adapter used for all currencies that do not register a custom one
+ * (Bitcoin, Litecoin, Dogecoin, Bitcoin Cash, etc.).
+ */
+export const defaultAdapter: ChainAdapter = {
+  id: "default",
+};
