+++ libs/coin-modules/coin-bitcoin/src/chain-adapters/zcash/operations.ts
@@ -1,0 +1,110 @@
+import { BigNumber } from "bignumber.js";
+import type { OperationType } from "@ledgerhq/types-live";
+import type { ShieldedTransaction } from "@ledgerhq/zcash-shielded/types";
+import { encodeOperationId } from "@ledgerhq/ledger-wallet-framework/operation";
+import type { BtcOperation } from "../../types";
+
+export const getTxType = (tx: ShieldedTransaction): OperationType => {
+  if (!tx.decryptedData) return "UNKNOWN";
+
+  const orchardOutputs = tx.decryptedData.orchard_outputs ?? [];
+  const saplingOutputs = tx.decryptedData.sapling_outputs ?? [];
+
+  let netDelta = new BigNumber(0);
+  for (const note of [...orchardOutputs, ...saplingOutputs]) {
+    if (note.transfer_type === "incoming") netDelta = netDelta.plus(note.amount);
+    else if (note.transfer_type === "outgoing") netDelta = netDelta.minus(note.amount);
+  }
+
+  if (netDelta.isGreaterThan(0)) {
+    return orchardOutputs.some(n => n.transfer_type === "incoming")
+      ? "SHIELDED_TX_ORCHARD_IN"
+      : "SHIELDED_TX_SAPLING_IN";
+  }
+  if (netDelta.isLessThan(0)) {
+    return orchardOutputs.some(n => n.transfer_type === "outgoing")
+      ? "SHIELDED_TX_ORCHARD_OUT"
+      : "SHIELDED_TX_SAPLING_OUT";
+  }
+  return "SHIELDED_TX_INTERNAL";
+};
+
+export function applyOutputDelta(
+  balance: BigNumber,
+  outputs: { transfer_type: string; amount: BigNumber }[],
+): BigNumber {
+  let result = balance;
+  for (const n of outputs) {
+    if (n.transfer_type === "incoming") result = result.plus(n.amount);
+    else if (n.transfer_type === "outgoing") result = result.minus(n.amount);
+  }
+  return result;
+}
+
+export function computeProtocolDeltas(transactions: ShieldedTransaction[]): {
+  deltaSapling: BigNumber;
+  deltaOrchard: BigNumber;
+} {
+  let deltaSapling = new BigNumber(0);
+  let deltaOrchard = new BigNumber(0);
+  for (const tx of transactions) {
+    deltaSapling = applyOutputDelta(deltaSapling, tx.decryptedData?.sapling_outputs ?? []);
+    deltaOrchard = applyOutputDelta(deltaOrchard, tx.decryptedData?.orchard_outputs ?? []);
+  }
+  return { deltaSapling, deltaOrchard };
+}
+
+export function computeOutgoingFees(transactions: ShieldedTransaction[]): BigNumber {
+  let fees = new BigNumber(0);
+  for (const tx of transactions) {
+    if (getTxType(tx).endsWith("_OUT")) {
+      fees = fees.plus(tx.fee ?? new BigNumber(0));
+    }
+  }
+  return fees;
+}
+
+/**
+ * Converts raw shielded transactions to BtcOperation format.
+ */
+export function convertShieldedTransactionsToOperations(
+  shieldedTransactions: ShieldedTransaction[],
+  accountId: string,
+): BtcOperation[] {
+  return shieldedTransactions.map(tx => {
+    const txType = getTxType(tx);
+    const allNotes = [
+      ...(tx.decryptedData?.orchard_outputs ?? []),
+      ...(tx.decryptedData?.sapling_outputs ?? []),
+    ];
+    let relevantTransferType: string | null = null;
+    if (txType.endsWith("_IN")) {
+      relevantTransferType = "incoming";
+    } else if (txType.endsWith("_OUT")) {
+      relevantTransferType = "outgoing";
+    }
+    const value = allNotes
+      .filter(n => n.transfer_type === relevantTransferType)
+      .reduce((sum, n) => sum.plus(n.amount), new BigNumber(0));
+
+    const operation: BtcOperation = {
+      id: encodeOperationId(accountId, tx.id, txType),
+      hash: tx.id,
+      accountId,
+      blockHash: tx.blockHash,
+      blockHeight: tx.blockHeight,
+      type: txType,
+      senders: [],
+      recipients: [],
+      date: new Date(tx.timestamp),
+      value,
+      fee: new BigNumber(tx.fee),
+      extra: {},
+      transactionSequenceNumber: new BigNumber(tx.blockHeight),
+      subOperations: [],
+      nftOperations: [],
+    };
+
+    return operation;
+  });
+}
